pyproject.toml
[tool.poetry]
name = "icart-backend"
version = "0.1.0"
description = ""
authors = ["arash3f <arash.alfooneh@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.103.1"
uvicorn = "^0.23.2"
alembic = "^1.12.0"
pydantic-extra-types = "^2.1.0"
pydantic-settings = "^2.0.3"
python-jose = "^3.3.0"
passlib = "^1.7.4"
minio = "^7.1.16"
asyncpg = "^0.28.0"
sqlalchemy = "^2.0.21"
python-multipart = "^0.0.6"
orjson = "^3.9.7"
requests = "^2.31.0"
kavenegar = "^1.1.2"
psycopg2-binary = "^2.9.8"
gunicorn = "^21.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
